refactor(tests): move build_labels tests back to test_collate_fns and delete test_label_utils
The label functions are back in vlm/collate_fns.py, so their tests
belong in test_collate_fns.py where they originally lived. Remove the
now-unnecessary test_label_utils.py file entirely.
